fix(auth): accept configured api key on admin /api endpoints
Unify api-key validation across admin /api/* and SAB /api surfaces so
both accept the configured api.key or FRONTEND_BACKEND_API_KEY.

Closes #242
