+++ backend/Api/SabControllers/SabApiController.cs
@@ -11,11 +11,11 @@ using NzbWebDAV.Api.SabControllers.GetStatus;
 using NzbWebDAV.Api.SabControllers.GetVersion;
 using NzbWebDAV.Api.SabControllers.RemoveFromHistory;
 using NzbWebDAV.Api.SabControllers.RemoveFromQueue;
+using NzbWebDAV.Auth;
 using NzbWebDAV.Config;
 using NzbWebDAV.Database;
 using NzbWebDAV.Extensions;
 using NzbWebDAV.Queue;
-using NzbWebDAV.Utils;
 using NzbWebDAV.Websocket;
 using Serilog;
 
@@ -118,17 +118,7 @@ public class SabApiController(
         public Task<IActionResult> HandleRequest()
         {
             if (RequiresAuthentication)
-            {
-                var apiKey = httpContext.GetRequestApiKey();
-                var isValidKey = apiKey?.IsAny(
-                    configManager.GetApiKey(),
-                    EnvironmentUtil.GetRequiredVariable("FRONTEND_BACKEND_API_KEY")
-                );
-                if (!isValidKey.HasValue)
-                    throw new UnauthorizedAccessException("API Key Required");
-                if (!isValidKey.Value)
-                    throw new UnauthorizedAccessException("API Key Incorrect");
-            }
+                ApiKeyValidator.Validate(httpContext, configManager);
 
             return Handle();
         }
